requirements.txt
joblib==1.0.1
tensorflow==2.5.0
Django==1.11.29
numpy==1.19.5
pandas==1.1.5
djangorestframework==3.12.4
rich==10.7.0
scikit_learn==0.24.2
